requirements.txt
# Data Manipulation
pandas>=2.0.0
numpy

# Database & ORM
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary 
supabase>=2.0.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Validation & Settings
pydantic>=2.0.0
pydantic-settings
python-dotenv

# Testing
pytest
httpx